Geez, lots of stuff here.
 * Got the main loop refactored to use cl-heap:priority-queue to manage getting objects to work at the right time.
 * pulled in the dice package to determine things randomly using terminology like "1d6 + 3"
 * got torpedos working (ran into a nasty bug because #(0 0) is not (vector 0 0))
 * finally got the outnumbered red team winning once in a while, armed with their torpedos
 * ships now have a list of available weapons, and choose what to fire randomly (and based on range/energy consumption)
 * added some new generics: is-alive-p, death
 * added some new classes: movable, torpedo
 * used cl-heap again to get a list of enemy ships in order of distance

Noticed that turning the frame-rate up to 30 makes the simulation speed up as more actors are destroyed.  I have a lot
of naive algorithms in here, it might be worth doing some refactoring to minimize uses of the distance formula, which
is redundantly calculated all over the place.

Fun stuff.